Split out main logic of DOM manipulation
Also started caching network responses (to avoid annoying Displate)
