+++ src/api.ts
@@ -1,7 +1,13 @@
 import { LimitedEdition } from "./model/limited-edition";
 
+// Let's not reload the data unless the user refreshes the page
+let cache: LimitedEdition[] = [];
+
 export async function fetchAllLimitedEditionData(): Promise<LimitedEdition[]> {
-  const response = await fetch("https://sapi.displate.com/artworks/limited");
-  const json = await response.json();
-  return json.data;
+  if (cache.length === 0) {
+    const response = await fetch("https://sapi.displate.com/artworks/limited");
+    const json = await response.json();
+    cache = json.data;
+  }
+  return cache;
 }
